fix(specs): remove unnecessary `as u8` cast on MAX_SIGNERS
MAX_SIGNERS is already u8, so `as u8` triggers clippy::unnecessary_cast
which fails CI lint with --deny=warnings.

+++ specs/shared/test_process_approve_checked_multisig.rs
@@ -17,10 +17,10 @@ fn test_process_approve_checked_multisig(
     #[cfg(feature = "assumptions")]
     {
         let multisig = get_multisig(&accounts[3]);
-        if multisig.m < 1 || multisig.m > MAX_SIGNERS as u8 {
+        if multisig.m < 1 || multisig.m > MAX_SIGNERS {
             return Ok(());
         }
-        if multisig.n < 1 || multisig.n > MAX_SIGNERS as u8 {
+        if multisig.n < 1 || multisig.n > MAX_SIGNERS {
             return Ok(());
         }
     }
